兼容JDK 17, 修复JDK 17 下用户保存报错的问题 !!!
