+++ erupt-extra/erupt-job/src/main/java/xyz/erupt/job/model/EruptMail.java
@@ -2,18 +2,12 @@ package xyz.erupt.job.model;
 
 import lombok.Getter;
 import lombok.Setter;
-import lombok.SneakyThrows;
-import org.apache.commons.lang3.StringUtils;
 import org.hibernate.annotations.Type;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.mail.javamail.JavaMailSenderImpl;
-import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.stereotype.Component;
 import xyz.erupt.annotation.Erupt;
 import xyz.erupt.annotation.EruptField;
 import xyz.erupt.annotation.EruptI18n;
 import xyz.erupt.annotation.config.EruptSmartSkipSerialize;
-import xyz.erupt.annotation.fun.DataProxy;
 import xyz.erupt.annotation.sub_erupt.Power;
 import xyz.erupt.annotation.sub_field.Edit;
 import xyz.erupt.annotation.sub_field.EditType;
@@ -22,16 +16,14 @@ import xyz.erupt.annotation.sub_field.sub_edit.BoolType;
 import xyz.erupt.annotation.sub_field.sub_edit.InputType;
 import xyz.erupt.annotation.sub_field.sub_edit.Search;
 import xyz.erupt.core.constant.RegexConst;
-import xyz.erupt.core.context.MetaContext;
-import xyz.erupt.core.util.Erupts;
+import xyz.erupt.job.model.data_proxy.EruptMailDataProxy;
 import xyz.erupt.jpa.model.BaseModel;
 
-import javax.mail.internet.MimeMessage;
-import javax.persistence.*;
-import java.nio.charset.StandardCharsets;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Lob;
+import javax.persistence.Table;
 import java.util.Date;
-import java.util.Objects;
-import java.util.Optional;
 
 /**
  * @author YuePeng
@@ -40,7 +32,7 @@ import java.util.Optional;
 @EruptI18n
 @Erupt(
         name = "发送邮件",
-        dataProxy = EruptMail.class,
+        dataProxy = EruptMailDataProxy.class,
         power = @Power(edit = false),
         orderBy = "id desc"
 )
@@ -49,7 +41,7 @@ import java.util.Optional;
 @Setter
 @Table(name = "e_job_mail")
 @Component
-public class EruptMail extends BaseModel implements DataProxy<EruptMail> {
+public class EruptMail extends BaseModel {
 
     @EruptField(
             views = @View(title = "接收人"),
@@ -98,30 +90,5 @@ public class EruptMail extends BaseModel implements DataProxy<EruptMail> {
     @EruptSmartSkipSerialize
     private String createBy;
 
-    @Transient
-    @Autowired(required = false)
-    private JavaMailSenderImpl javaMailSender;
 
-    @SneakyThrows
-    @Override
-    public void beforeAdd(EruptMail eruptMail) {
-        Erupts.requireNonNull(javaMailSender, "Sending mailbox not configured");
-        MimeMessage mimeMessage = javaMailSender.createMimeMessage();
-        MimeMessageHelper helper = new MimeMessageHelper(mimeMessage, true, StandardCharsets.UTF_8.name());
-        eruptMail.setCreateBy(MetaContext.getUser().getName());
-        eruptMail.setCreateTime(new Date());
-        helper.setSubject(eruptMail.getSubject());
-        helper.setTo(eruptMail.getRecipient());
-        helper.setFrom(Objects.requireNonNull(javaMailSender.getUsername()));
-        if (StringUtils.isNotBlank(eruptMail.getCc())) helper.setCc(eruptMail.getCc().split("\\|"));
-        helper.setText(eruptMail.getContent(), true);
-        try {
-            javaMailSender.send(mimeMessage);
-            eruptMail.setStatus(true);
-        } catch (Exception e) {
-            e.printStackTrace();
-            eruptMail.setStatus(false);
-            Optional.ofNullable(e.toString()).ifPresent(it -> eruptMail.setErrorInfo(it.substring(0, 5000)));
-        }
-    }
 }
